ColorsActivity,java, FamilyActivity,java, NumbersActivity.java I added all the images for the words and I associated the words with the right image.
Word.java
Because Phrases are not associated with any image, a check was needed. The hasImage() will return true if the word had an image.

WordAdapter.java
This class created the views. If hasImage() returns true, it will add the image to the view and set the visibility to VISIBLE. Otherwise it will set the visibility to GONE. setVisibility() method was needed in both cases because the iconView remembers the last settings.