Implementing Fragments.
This task was done by creating a Fragment for each Activity(except the Main Activity). The code from the Fragment classes is similar with the one from the Activities. The differences consist in the:

onCreateView() - Within the Activity’s onCreate() method, we can call setContentView() to set the layout for the activity. In the Fragment, we need to inflate the view from the XML layout resource ID and return that view in the onCreateView() method .

onStop() - here I used releaseMediaPlayer()

In the Activity classes:

onCreate() - is the only code left in these classes. This method contains a setContentView() and a getSupportFragmentManager() which will replace the container xml with the Fragment.